requirements.txt
PyYAML>=6.0
Pillow>=9.0.0
jinja2>=3.0
llama-cpp-python>=0.1.64
transformers>=4.0.0

# Music Analysis Libraries
librosa>=0.10.0
numpy>=1.21.0
scipy>=1.7.0
essentia-tensorflow>=2.1b6.dev1034
madmom>=0.16.1
torch>=1.9.0
muq>=0.1.0

